# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/m-role": {
            "get": {
                "description": "Get Page MRole",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mRole"
                ],
                "summary": "MRolePage",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "default": "0",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "5",
                        "description": "size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "global filter",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Update MRole",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mRole"
                ],
                "summary": "MRoleUpdate",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "description": "Add MRoleRequest",
                        "name": "mRole",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/m_role.MRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Create MRole",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mRole"
                ],
                "summary": "MRoleCreate",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "description": "Add MRoleRequest",
                        "name": "mRole",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/m_role.MRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/m-role/{id}": {
            "get": {
                "description": "Get MRole by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mRole"
                ],
                "summary": "MRoleIndex",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "MRole id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete MRole by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mRole"
                ],
                "summary": "MRoleDelete",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "MRole id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "m_role.MRoleRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 20
                },
                "id": {
                    "type": "integer"
                },
                "isDelete": {
                    "type": "boolean"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "path": {
                    "type": "string",
                    "example": "/v1/m_biodata"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                }
            }
        }
    }
}